Allow some formatting in comments
These changes allow some formatting inside the comments.  Possibly not
the best way to go about it, but it seems to work ok.  Markup is limited
to things that fit on a single line, but can include bolding, images,
etc.  Links wouldn't be very useful - no way to move the mouse over to
click them.